requirements.txt
mecab-python3
jaconv
numpy
unidic-lite
Misskey.py